Fix signed/unsigned comparison for write() return in upload path
The download path at line 345 correctly casts the expected size to
ssize_t before comparing with write()'s return value. The upload
path at line 354 compared write()'s ssize_t return with size*nmemb
(size_t) without the cast, creating a signed/unsigned mismatch.
Add the same (ssize_t) cast for consistency and correctness.

+++ modules/curl/module.c
@@ -351,7 +351,7 @@ static size_t curl_writefunction_cb(void *ptr, size_t size, size_t nmemb, void *
 			return size*nmemb;
 
 		g_debug("session %p file %i", session, session->action.upload.file->fd);
-		if( write(session->action.upload.file->fd, ptr, size*nmemb) != size*nmemb)
+		if( write(session->action.upload.file->fd, ptr, size*nmemb) != (ssize_t)(size*nmemb))
 			return 0;
 	}
